+++ TheGame - Server/src/thegame/com/Game/Objects/ToolType.java
@@ -14,7 +14,7 @@ public class ToolType implements Serializable{
     public float speed;
     public float range;
     public int reqLvl;
-    public int kb;
+    public float kb;
     public float height;
     public float width;
     public toolType type; 
@@ -33,7 +33,7 @@ public class ToolType implements Serializable{
      * @param height
      * @param width
      */
-    public ToolType(String name, int strength, float speed, float range, int reqLvl, toolType type, int kb, float height, float width) {
+    public ToolType(String name, int strength, float speed, float range, int reqLvl, toolType type, float kb, float height, float width) {
         this.name = name;
         this.strength = strength;
         this.speed = speed;
